requirements.txt
asgiref==3.3.4
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.2
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-import-export==2.4.0
django-pandas==0.6.4
django-storages==1.12.3
django-tables2==2.3.3
djangorestframework==3.12.2
et-xmlfile==1.0.1
google-api-core==1.31.5
google-api-python-client==1.12.8
google-auth==1.35.0
google-auth-httplib2==0.0.4
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.52.0
greenlet==1.0.0
httplib2==0.18.1
idna==2.10
importlib-metadata==4.0.1
jdcal==1.4.1
json2table==1.1.5
MarkupPy==1.14
mysqlclient==2.0.1
numpy==1.20.2
odfpy==1.4.1
openpyxl==3.0.5
packaging==21.3
pandas==1.2.4
pdfkit==1.0.0
Pillow==8.4.0
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.1
pyparsing==3.0.6
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
reportlab==3.6.3
requests==2.25.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.4.11
sqlparse==0.4.1
tablib==2.0.0
twilio==6.57.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.2.0
xlrd==1.2.0
XlsxWriter==1.4.0
xlwt==1.3.0
xmltodict==0.12.0
zipp==3.4.1
